Expanded Game.py
the game window now a menu bar and the load menu. The menu bar has multiple cascades which are used to load, save, exit and change the buy mode. Added some comments to separate the types of methods (was getting too messy). Also added a getResources() methods to get all the .png's