Changes
Added labels to display the amount of things you own. Removed the status bar text for the same. Added an update method that updates the text in all the labels on screen. Changed the name of the buy(name) method to addToPlayer(name). Reordered the buyable dict to match with the buttons. Added an empty.png which is an empty png. It is just so that the height option works in pixels and not lines for the amount labels.